testing/mochitest/tests/test_syntheticMouseEvent.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
    <title>Test for Bug 633602</title>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
    <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
	<div id="content">
        <script type="application/javascript">
            var point1 = false;
            var point2 = false;
            var point3 = false;
            var pointer = navigator.pointer;
        </script>
        <map name="map" id="map">
            <area shape="rect" coords="1,1,1,1" onclick="point1 = true;" alt="point 1"/>
            <area shape="rect" coords="2,1,2,1" onclick="point2 = true;" alt="point 2"/>
            <area shape="rect" coords="3,1,3,1" onclick="point3 = true;" alt="point 3"/>
        </map>
        <img id="image" src="image.png" usemap="#map" width="200" height="200" alt="Synthetic Mouse Event Test Image"/>
	</div>
    <pre id="test">
        <script type="application/javascript">
            /** Test for Bug 633602 **/
            /** Test to see if Synthetic Mouse behave the same regardless of mouse lock state**/
            SimpleTest.waitForExplicitFinish();
            SimpleTest.waitForFocus(function () {
                var image = document.getElementById("image");
                synthesizeMouse(image, 1, 1, {});
                synthesizeMouse(image, 2, 1, {});
                synthesizeMouse(image, 3, 1, {});
                /* Resetting the test for mouse lock */
                var unlock1 = point1;
                point1 = false;
                var unlock2 = point2;
                point2 = false;
                var unlock3 = point3
                point3 = false;
                /* Run the same test with mouse locked */
                pointer.lock();
                synthesizeMouse(image, 1, 1, {});
                synthesizeMouse(image, 2, 1, {});
                synthesizeMouse(image, 3, 1, {});
                is(unlock1, point1, "Point 1 click failed");
                is(unlock1, point1, "Point 2 click failed");
                is(unlock1, point1, "Point 3 click failed");
                SimpleTest.finish();
            });
        </script>
    </pre>
</body>
</html>
